Fixed a warning when compiling for 64 bit

+++ tuxcap/chipmunk/chipmunk.h
@@ -51,6 +51,8 @@ cpfmin(cpFloat a, cpFloat b)
 	#endif
 #endif
 
+#include <inttypes.h>
+
 #include "cpVect.h"
 #include "cpBB.h"
 #include "cpBody.h"
@@ -69,7 +71,7 @@ cpfmin(cpFloat a, cpFloat b)
 #include "cpSpace.h"
 
 #define CP_HASH_COEF (3344921057ul)
-#define CP_HASH_PAIR(A, B) ((unsigned int)(A)*CP_HASH_COEF ^ (unsigned int)(B)*CP_HASH_COEF)
+#define CP_HASH_PAIR(A, B) ((uintptr_t)(A)*CP_HASH_COEF ^ (uintptr_t)(B)*CP_HASH_COEF)
 
 void cpInitChipmunk(void);
